+++ Sources/PortalSwift/Portal.swift
@@ -46,6 +46,7 @@ public final class Portal: PortalProtocol {
   /// Access to trading-related functionality.
   public lazy var trading: Trading = .init(
     api: self.api,
+    portal: self,
     signAndSendTransaction: { [weak self] transaction, chainId in
       guard let self else {
         throw PortalClassError.clientNotAvailable
@@ -2529,6 +2530,8 @@ public final class Portal: PortalProtocol {
 }
 
 extension Portal: EvmAccountTypePortalDependency {}
+
+extension Portal: ZeroXPortalDependency {}
 
 extension Portal: YieldXyzPortalDependency {}